1465: moved all math dependencies into the src/stan/math folder; stan::memory namespace changed to stan::math
