+++ .travis.yml
@@ -16,12 +16,12 @@ env:
   - TESTFOLDER=src/test/unit/interface
   - TESTFOLDER=src/test/unit/io
   - TESTFOLDER=src/test/unit/lang
+  - TESTFOLDER=src/test/unit/math/memory
   - TESTFOLDER=src/test/unit/math/prim
   - TESTFOLDER=src/test/unit/math/rev
   - TESTFOLDER=src/test/unit/math/fwd
  # - TESTFOLDER=src/test/unit/math/mix  ## commenting this out due to segfault on travis
   - TESTFOLDER=src/test/unit/mcmc
-  - TESTFOLDER=src/test/unit/memory
   - TESTFOLDER=src/test/unit/model
   - TESTFOLDER=src/test/unit/optimization
   - TESTFOLDER=src/test/unit/services
