+++ src/doxygen/stan.dox
@@ -27,11 +27,6 @@
  */
 
 /**
- * @namespace stan::memory 
- * (Expert) Allocate, free, and manage underlying system memory.
- */
-
-/**
  * @namespace stan::model
  * For compiling models.
  */
