+++ src/stan/math/rev/core/autodiffstackstorage.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_REV_CORE_AUTODIFFSTACKSTORAGE_HPP
 #define STAN_MATH_REV_CORE_AUTODIFFSTACKSTORAGE_HPP
 
-#include <stan/memory/stack_alloc.hpp>
+#include <stan/math/memory/stack_alloc.hpp>
 #include <vector>
 
 namespace stan {
@@ -13,7 +13,7 @@ namespace stan {
       static std::vector<ChainableT*> var_stack_;
       static std::vector<ChainableT*> var_nochain_stack_;
       static std::vector<ChainableAllocT*> var_alloc_stack_;
-      static memory::stack_alloc memalloc_;
+      static stack_alloc memalloc_;
 
       // nested positions
       static std::vector<size_t> nested_var_stack_sizes_;
@@ -34,7 +34,7 @@ namespace stan {
     AutodiffStackStorage<ChainableT, ChainableAllocT>::var_alloc_stack_;
 
     template<typename ChainableT, typename ChainableAllocT>
-    memory::stack_alloc
+    stack_alloc
     AutodiffStackStorage<ChainableT, ChainableAllocT>::memalloc_;
 
     template<typename ChainableT, typename ChainableAllocT>
