--- src/stan/memory.hpp
@@ -1,6 +1,0 @@
-#ifndef STAN_MEMORY_HPP
-#define STAN_MEMORY_HPP
-
-#include <stan/memory/stack_alloc.hpp>
-
-#endif
