+++ src/test/unit/math/rev/mat/fun/log_softmax_test.cpp
@@ -14,7 +14,7 @@
 
 TEST(AgradRevMatrix,logSoftmaxLeak) {
   // FIXME: very brittle test depending on unrelated constants of 
-  //        block sizes/growth in stan::memory::stack_alloc
+  //        block sizes/growth in stan::math::stack_alloc
   using stan::math::log_softmax;
   using stan::math::log_softmax;
   using Eigen::Matrix;
